requirements.txt
captcha==0.4
configparser==3.5.0
mysql-connector-python==8.0.18
Pillow==9.0.1
protobuf==3.19.4
pyaes==1.6.1
Pyrogram==1.4.9
PySocks==1.7.1
TgCrypto==1.2.3
